Converted ExpansionTableDemo into a live code example on the documentation website
